refactor(frontend): moved user to own file
